requirements.txt
asgiref==3.7.2
Django==4.2.3
factory_boy==3.3.3
Faker==37.6.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2025.2
